+++ lib/cronex/job.ex
@@ -62,7 +62,7 @@ defmodule Cronex.Job do
 
   # Every week job, check time and day of the week
   defp is_time({minute, hour, :*, :*, day_week}) do
-    current_date_time.minute == minute and current_date_time.hour == hour and current_date_time.day_of_the_week == day_week
+    current_date_time.minute == minute and current_date_time.hour == hour and current_date_time.day_of_week == day_week
   end
 
   # Every month job, check day and time of job