+++ lib/cronex/parser.ex
@@ -21,10 +21,20 @@ defmodule Cronex.Parser do
 
       iex> Cronex.Parser.parse_frequency(:wednesday, "12:00")
       {0, 12, :*, :*, 3}
+
+      iex> Cronex.Parser.parse_frequency(:non_existing_day)
+      :invalid
+
+      iex> Cronex.Parser.parse_frequency(:monday, "invalid time")
+      :invalid
   """
   def parse_frequency(frequency, time \\ "00:00") do
-    {hour, minute} = parse_time(time)
-
+    parsed_time = parse_time(time)
+    do_parse_frequency(frequency, parsed_time)
+  end
+         
+  defp do_parse_frequency(_, :invalid), do: :invalid
+  defp do_parse_frequency(frequency, {hour, minute}) do
     cond do
       frequency == :minute ->
         {:*, :*, :*, :*, :*}
@@ -44,13 +54,19 @@ defmodule Cronex.Parser do
       frequency in @days_of_week ->
         day_of_week = Enum.find_index(@days_of_week, &(&1 == frequency)) + 1
         {minute, hour, :*, :*, day_of_week}
+
+      true -> :invalid 
     end
   end
 
   defp parse_time(time) when is_bitstring(time) do
-    time 
-    |> String.split(":")
-    |> Enum.map(&String.to_integer/1)
-    |> List.to_tuple
+    try do
+      time 
+      |> String.split(":")
+      |> Enum.map(&String.to_integer/1)
+      |> List.to_tuple
+    rescue
+      _ -> :invalid
+    end
   end
 end